reader.bmx: Moved code for encoding/decoding JSON strings to jsonstrings.bmx, as well as JHexCharToByte (no longer used by JReader specifically).
characters.bmx:
Added - contains constants that are included in jsonstring.bmx, and possibly reader.bmx later.  Should be included in a private section.

jsonstrings.bmx:
Optimization- EncodeJSONString (formerly JString.ToString) rewritten to use a temporary buffer instead of roughly 32 calls to Replace - should reduce the amount of memory allocated for the function by about 94%, and if the encoded string would not be different, there should be zero allocations.

jobjecttypes.bmx:
Added code to determine the type of an object as it pertains to JSON.  JInvalidType should be noted, as this will be returned for objects that don't match specific types.